+++ app/Http/Controllers/ProjectController.php
@@ -5,22 +5,33 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Project;
 use App\Models;
+use App\User;
 use App\Http\Requests\ProjectRequest;
 use App\Http\Requests\ProjectUpdateRequest;
+use Illuminate\Support\Facades\DB;
 
 class ProjectController
 {
-    //
+    
     public function index(Request $request)
     {
+        $projects = $request->user()->projects();
+
         return response()->json([
-             'success' => true,
-             'projects' => Project::with(['epics', 'epics.user_stories'])->get()
+            'success' => true,
+            'projects' => $projects->with(['epics', 'epics.user_stories'])->get()
         ]);
     }
 
-    public function show(Project $project)
+    public function show(Request $request, Project $project)
     {
+        if(!$request->user()->has_basic_permissions_to_project($project)){
+            return response()->json([
+                'success' => false,
+                'message' => 'Not authorized'
+             ], 403);
+        }
+
         return response()->json([
             'success' => true,
             'project' => $project->load(['epics', 'epics.user_stories'])
@@ -36,8 +47,6 @@ class ProjectController
     {
         $project = new Project($request->all());
         $project->save();
-
-        
         $project->users()->attach($request->userId , array('role_id' => 1));
 
         return response()->json([
@@ -47,8 +56,14 @@ class ProjectController
         
     }
 
-    public function edit(Project $project)
+    public function edit(Request $request, Project $project)
     {
+        if(!$request->user()->has_basic_permissions_to_project($project)){
+            return response()->json([
+                'success' => false,
+                'message' => 'Not authorized'
+             ], 403);
+        }
         
         return response()->json([
             'success' => true,
@@ -58,6 +73,13 @@ class ProjectController
 
     public function update(ProjectUpdateRequest $request, Project $project){
 
+        if(!$request->user()->has_basic_permissions_to_project($project)){
+            return response()->json([
+                'success' => false,
+                'message' => 'Not authorized'
+             ], 403);
+        }
+
         $project->update($request->all());
         
         return response()->json([
@@ -66,8 +88,15 @@ class ProjectController
         ]);
     }
 
-    public function delete(Project $project)
+    public function delete(Request $request, Project $project)
     {
+        if(!$request->user()->has_full_permissions_to_project($project)){
+            return response()->json([
+                'success' => false,
+                'message' => 'Not authorized'
+             ], 403);
+        }
+
         $project->delete();
 
         return response()->json([
